Add dynamic content and interactivity to admin panel
Refactor `admin.js` to introduce event listeners for dynamic elements, enhancing user interactivity for knowledge, organization, and data source management. Update `admin.html` to support new modal dialogs for managing files and links, improving the UI/UX experience.

+++ frontend/admin.js
@@ -4,8 +4,90 @@ interceptor.overrideFetch();
 const apiUrl = interceptor.apiUrl;
 
 window.onload = async function() {
-    await getAgentStatus();
-    await getOrganization();
+    getAgentStatus();
+    getOrganization();
+
+    document.getElementById("showStaticContainer").addEventListener('click', function(e) {
+        const staticContainer = document.getElementById('staticContainer');
+        const webContainer = document.getElementById('webContainer');
+        staticContainer.style.display = 'block';
+        webContainer.style.display = 'none';
+
+        document.getElementById('showStaticContainerLi').classList.add("active");
+        document.getElementById('showWebContainerLi').classList.remove("active");
+    });
+
+    document.getElementById("showWebContainer").addEventListener('click', function(e) {
+        const staticContainer = document.getElementById('staticContainer');
+        const webContainer = document.getElementById('webContainer');
+        staticContainer.style.display = 'none';
+        webContainer.style.display = 'block';
+
+        document.getElementById('showStaticContainerLi').classList.remove("active");
+        document.getElementById('showWebContainerLi').classList.add("active");
+    });
+
+    document.getElementById("enableKnowledge").addEventListener('click', function(e) {
+        enableKnowledge();
+    });
+
+    document.getElementById("disableKnowledge").addEventListener('click', function(e) {
+        disableKnowledge();
+    });
+
+    $('#staticModal').on('shown.bs.modal', function (e) {
+        fillStaticModal();
+    });
+
+    $('#webModal').on('shown.bs.modal', function (e) {
+        fillWebModal();
+    });
+
+    document.getElementById("orgDescription").addEventListener('change', function(e) {
+        document.getElementById('updateOrgButton').disabled = false;
+    });
+    document.getElementById("orgName").addEventListener('change', function(e) {
+        document.getElementById('updateOrgButton').disabled = false;
+    });
+
+    document.getElementById('updateOrgButton').addEventListener('click', function(e) {
+        updateProfile();
+    });
+
+    document.getElementById('updateOrgButton').disabled = true;
+
+    document.getElementById('inputStaticFileUpload').addEventListener('change', function(event) {
+        var fileName = event.target.files[0].name;
+        var label = document.querySelector('label[for="inputStaticFileUpload"]');
+        label.textContent = fileName;
+    });
+
+    document.getElementById('uploadFile').addEventListener('click', function() {
+        var input = document.getElementById('inputStaticFileUpload');
+        var file = input.files[0];
+
+        if (file) {
+            uploadFile(file);
+        } else {
+            alert('Please select a file to upload.');
+        }
+    });
+
+    document.getElementById('updateWebLinksButton').addEventListener('click', function(e) {
+        updateWebLinks();
+    });
+
+    document.getElementById('syncWebDataSource').addEventListener('click', function() {
+        syncWebDataSource();
+    });
+    document.getElementById('syncStaticDataSource').addEventListener('click', function() {
+        syncStaticDataSource();
+    });
+
+    document.getElementById('deleteWebDataSource').addEventListener('click', function() {
+        deleteWebDataSource();
+    });
+
 };
 
 
@@ -40,7 +122,9 @@ async function disableKnowledge() {
     if (!response.ok) {
         const result = await response.json();
         window.alert(result.message);
+        return
     }
+    location.reload(true);
 }
 
 async function enableKnowledge() {
@@ -48,26 +132,80 @@ async function enableKnowledge() {
     if (!response.ok) {
         const result = await response.json();
         window.alert(result.message);
+        return
     }
+    location.reload(true);
 }
 
-async function deleteKnowledgeFile() {
-    const response = await fetch(`${apiUrl}/api/admin/knowledge/static/delete-file`, { method: 'DELETE' });
-    if (!response.ok) {
-        const result = await response.json();
-        window.alert(result.message);
-    }
+async function syncStaticDataSource() {
+    const url = '/api/admin/knowledge/static/sync';
+    await syncDataSource(url);
 }
 
-async function uploadFile() {
-    const response = await fetch(`${apiUrl}/api/admin/knowledge/static/get-upload-link`);
+async function syncWebDataSource() {
+    const url = '/api/admin/knowledge/web/sync';
+    await syncDataSource(url);
+}
+
+async function syncDataSource(url) {
+    const response = await fetch(`${apiUrl}${url}`, { method: 'PUT' });
     if (!response.ok) {
         const result = await response.json();
         window.alert(result.message);
         return
     }
+    location.reload(true);
+}
+
+async function deleteKnowledgeFile(file) {
+    if (!confirm(`${file} will be deleted permanently. Are you sure?`)) {
+      return
+    }
+    const response = await fetch(
+        `${apiUrl}/api/admin/knowledge/static/delete-file?fileName=${file}`,
+        {
+            method: 'DELETE',
+            headers: {
+                "Content-Type": "application/json"
+            }
+        }
+    );
+    if (!response.ok) {
+        const result = await response.json();
+        window.alert(result.message);
+        throw new Error(result.message);
+    }
+}
+
+async function uploadFile(file) {
+    const response = await fetch(`${apiUrl}/api/admin/knowledge/static/get-upload-link?fileName=${file.name}`);
+    if (!response.ok) {
+        const result = await response.json();
+        window.alert(result.message);
+        throw new Error(result.message);
+    }
     const result = await response.json();
-//    TODO
+    const uploadUrl = result['presignedUrl'];
+
+    var xhr = new XMLHttpRequest();
+    xhr.open('PUT', uploadUrl, true);
+    xhr.setRequestHeader('Content-Type', file.type);
+    xhr.onload = () => {
+      if (xhr.status === 200) {
+        console.log('File uploaded successfully');
+        fillStaticModal();
+      }
+    };
+    xhr.onerror = () => {
+      console.log('File upload failed');
+    };
+    xhr.send(file);
+    xhr.upload.onprogress = (event) => {
+      if (event.lengthComputable) {
+        var percent = Math.round((event.loaded / event.total) * 100)
+        console.log(percent);
+      }
+    };
 }
 
 async function listKnowledgeFiles() {
@@ -75,37 +213,38 @@ async function listKnowledgeFiles() {
     const result = await response.json();
     if (!response.ok) {
         window.alert(result.message);
-        return
+        throw new Error(result.message);
     }
-//    TODO
+    return result;
 }
 
-async function syncKnowledge() {
-    const response = await fetch(`${apiUrl}/api/admin/knowledge/static/sync`, { method: 'POST' });
+async function updateWebLinks() {
+    const urlInputs = document.querySelectorAll('[data-web-data-source-url]');
+    let urls = [];
+    urlInputs.forEach((i) => {
+      urls.push(i.value);
+    });
+    const response = await fetch(
+        `${apiUrl}/api/admin/knowledge/web/update`,
+        {
+            method: 'POST',
+            body: JSON.stringify({
+                urls: urls
+            })
+        }
+    );
     if (!response.ok) {
-        const result = await response.json();
         window.alert(result.message);
+        throw new Error(result.message);
     }
+    fillWebModal();
 }
 
-async function addWebLink() {
-    const response = await fetch(`${apiUrl}/api/admin/knowledge/web/add-link`, { method: 'POST' });
-    if (!response.ok) {
-        const result = await response.json();
-        window.alert(result.message);
+async function deleteWebDataSource() {
+    if (!confirm(`This action is irreversible and all your urls are going to be deleted. Are you sure?`)) {
+      return
     }
-}
-
-async function deleteWeb() {
     const response = await fetch(`${apiUrl}/api/admin/knowledge/web/delete`, { method: 'DELETE' });
-    if (!response.ok) {
-        const result = await response.json();
-        window.alert(result.message);
-    }
-}
-
-async function deleteWebLink() {
-    const response = await fetch(`${apiUrl}/api/admin/knowledge/web/delete-link`, { method: 'DELETE' });
     if (!response.ok) {
         const result = await response.json();
         window.alert(result.message);
@@ -114,23 +253,25 @@ async function deleteWebLink() {
 
 async function listWebLinks() {
     const response = await fetch(`${apiUrl}/api/admin/knowledge/web/list`);
-    const result = await response.json();
-    if (!response.ok) {
-        window.alert(result.message);
-        return
-    }
-//    TODO
-}
-
-async function syncWeb() {
-    const response = await fetch(`${apiUrl}/api/admin/knowledge/web/sync`, { method: 'POST' });
-    if (!response.ok) {
+    let urls;
+    if (!response.ok && response.status == 404) {
+        urls = [];
+    } else if (!response.ok) {
         const result = await response.json();
         window.alert(result.message);
+        throw new Error(result.message);
+    } else {
+        const result = await response.json();
+        urls = result['urls'];
     }
+
+    return urls;
 }
 
 async function deleteProfile() {
+    if (!confirm(`Your profile is going to be deleted. This action is irreversible. Are you sure?`)) {
+      return
+    }
     const response = await fetch(`${apiUrl}/api/admin/organization/delete-profile`, { method: 'DELETE' });
     if (!response.ok) {
         const result = await response.json();
@@ -139,9 +280,104 @@ async function deleteProfile() {
 }
 
 async function updateProfile() {
-    const response = await fetch(`${apiUrl}/api/admin/organization/update`, { method: 'POST' });
+    const description = document.getElementById("orgDescription").value;
+    const name = document.getElementById("orgName").value;
+    const response = await fetch(
+        `${apiUrl}/api/admin/organization/update`,
+        {
+            method: 'POST',
+            body: JSON.stringify({
+                organization_name: name,
+                organization_description: description
+            })
+        }
+    );
     if (!response.ok) {
         const result = await response.json();
         window.alert(result.message);
+        throw new Error(result.message);
     }
+    location.reload(true);
+}
+
+function getUrlInput(text) {
+    const div1 = document.createElement("div");
+    div1.classList.add("input-group", "mb-3");
+    const input = document.createElement("input");
+    input.classList.add("form-control");
+    input.type = "text";
+    input.value = text;
+    input.setAttribute("data-web-data-source-url", true);
+    input.addEventListener('change', function(e) {
+        document.getElementById('updateWebLinksButton').disabled = false;
+    });
+    const div2 = document.createElement("div");
+    div1.classList.add("input-group-prepend");
+    const button = document.createElement("button");
+    button.type = "button";
+    button.classList.add("btn", "btn-outline-secondary");
+    button.innerText = 'Remove';
+    button.addEventListener('click', function(e) {
+        div1.remove();
+        document.getElementById('updateWebLinksButton').disabled = false;
+    });
+    div2.appendChild(button);
+    div1.appendChild(input);
+    div1.appendChild(div2);
+
+    return div1;
+}
+
+async function fillWebModal() {
+    const result = await listWebLinks();
+    const container = document.getElementById("staticWebListContainer");
+    const ul = document.createElement("ul");
+    const mainDiv = document.createElement("div");
+    mainDiv.id = 'staticWebUrlsList'
+    let div;
+    container.innerHTML = '';
+    for (const url of result) {
+        div = getUrlInput(url);
+        mainDiv.appendChild(div);
+    }
+    container.appendChild(mainDiv);
+    div = document.createElement("div");
+    div.style = "text-align: center; justify-content: center; align-items: center";
+    const addButton = document.createElement("button");
+    addButton.type = "button";
+    addButton.classList.add("btn", "btn-outline-secondary", "mb-3");
+    addButton.innerText = "+";
+    addButton.addEventListener('click', function(e) {
+        document.getElementById("staticWebUrlsList").appendChild(getUrlInput(''));
+    });
+    div.appendChild(addButton)
+    container.appendChild(div);
+
+    document.getElementById('updateWebLinksButton').disabled = true;
+}
+
+async function fillStaticModal() {
+    const result = await listKnowledgeFiles();
+    const container = document.getElementById('staticFilesListContainer');
+    const ul = document.createElement("ul");
+    let li;
+    let span;
+    let button;
+    for (const file of result['files']) {
+        li = document.createElement("li");
+        span = document.createElement("span");
+        button = document.createElement("button");
+        span.innerText = file;
+        button.innerText = 'X';
+        button.addEventListener('click', function(e) {
+            deleteKnowledgeFile(file);
+        });
+        button.style.float = 'right';
+        button.style.marginRight = '5%';
+        li.appendChild(span);
+        li.appendChild(button);
+        ul.appendChild(li);
+    }
+    container.innerHTML = '';
+    container.appendChild(ul);
 }
